requirements.txt
tweepy==3.9.0
kafka-python==2.0.1
pyspark==2.4.6